Continue processing SIP messaging even if push notification sent
This fixes an issue so devices that remain registered after sleeping
can still be woke by a push. Also, if it happens that their socket is
still open, as happens for a short time during background fetch on iOS,
they can answer the SIP signalling on that socket and handle the call.