Ensure $sht(auth_cache=>$Au) != 0 when comparing with authn_perm_err
When using x-token-reg in pusher to do fast reg, $var(password) is set
to $null, which results in a value of 0. This value is stored in the
auth_cache and leads to `rval_get_int(): automatic string to int
conversion for "authn_perm_err" failed`. The fallback behaviour makes
the if condition true, which isn't what is desired.